spec: view modes and actions displays 'view' by default - READ BELOW!
This could be considered a canonical example about why you should TDD.
The test-pain is so great that makes this whole testing thing completely worthless.
The pain is buried in indirection and abstraction (beforeEach), so the spec is readable.
58 lines of code to test which view is displayed by default.... WTF!!!!!!
The underlying problem?
    1 - Take a look at the constructor in app/js/app/views/note.js --> 18 lines
        This should be considered an anti-pattern...  SRP anyone
    2 - I'm not sure if this problem is a Backbone design problem
        or simply a result of a very poor designed app.
    3 - Another thing to consider is that maybe this kind of behaviour
        should be tested (aka driven-designed) anywhere else.
        Maybe from an acceptance test (Cucumber) ???
Anyway, this kind of test and corresponding constructor
should never be written in production.